+++ finerplan/finerplan.py
@@ -9,10 +9,18 @@ app.config.from_object(__name__) # load config from this file
 app.config.update(dict(
     NAME = os.getenv("LOGNAME").capitalize(),
     DATABASE = os.path.join(app.root_path, 'finerplan.db'),
-    CREDIT_CLOSING = 11, # Insert here the day of month when the credit card invoice closes
-    CREDIT_PAYMENT = 25, # Insert here the day of month when the credit card invoice is paid
+    CREDIT_CLOSING = 11, # Day of month when the credit card invoice closes
+    CREDIT_PAYMENT = 25, # Day of month when the credit card invoice is paid
     SECRET_KEY = (os.environ.get('SECRET_KEY') or "you-will-never-guess"),
 ))
+
+# Defines common words used in the forms
+form_words = {'earnings': "Receita",
+              'expenses': "Gasto",
+              'assets': "Investimento",
+              'cash': "Dinheiro",
+              'credit': "Crédito",
+              'outsourced': "Terceiros"}
 
 # Creates the tables if they don't exist.
 con = sqlite3.connect(app.config['DATABASE'],  check_same_thread=False)