+++ finerplan/routes.py
@@ -1,38 +1,39 @@
 from flask import render_template, flash
+from .forms import AddTransactionForm, TestForm
+
+from finerplan import sql, dates, reports
 from .finerplan import app
-from .forms import AddTransactionForm
-from finerplan import sql, dates
 
 @app.route('/', methods=['GET', 'POST'])
 @app.route('/index', methods=['GET', 'POST'])
 def index():
     form = AddTransactionForm()
     if form.validate_on_submit():
-        if form.transaction.data == "Receita":
-            if sql.insert_earning(form):
-                flash("The earning wasn't inserted correctly")
-            else:
-                flash("Added new earning '{}', occurred in {} with the value of R$ {}".format(form.description.data, form.date.data, form.value.data))
-        elif form.transaction.data == "Investimento":
-            if sql.insert_investment(form):
-                flash("The investment wasn't inserted correctly")
-            else:
-                flash("Added new investment '{}', occurred in {} with the value of R$ {}".format(form.description.data, form.date.data, form.value.data))
+        flash("Form validated")
+        if form.transaction.data:
+            flash("Inserted new entry in {} table".format(form.transaction.data))
+            err = sql.insert_entry(form)
+        if err:
+            flash("The new entry wasn't inserted correctly")
         else:
-            if sql.insert_expense(form):
-                flash("The expense wasn't inserted correctly")
-            else:
-                flash("Added new expense '{}', occurred in {} with the value of R$ {}".format(form.description.data, form.date.data, form.value.data))
-                
-    report = {'name': app.config['NAME'], 
-              'eom': dates.s_eom, 
-              'balance': sql.balance_eom(local=True), 
-              'income': sql.income_ocm(local=True), 
-              'savings': sql.savings_ocm(local=True), 
-              'sav_rate': sql.savings_rate(), 
-              'credit': sql.credit_payment(local=True),
-              'expenses':sql.last_expenses(), 
-              'earnings':sql.last_earnings(), 
-              'investments':sql.last_investments(), 
-              'categorias': [ elem[1] for elem in sql.generate_categories()[0] ]}
-    return render_template('overview.html', title='Overview', form=form,  report=report)
+            flash("Successfully added new {}".format(form.transaction.data.lower()))
+    tables = {'expenses':sql.last_expenses(),
+              'earnings':sql.last_earnings(),
+              'investments':sql.last_investments()}
+    basic_report = reports.basic()
+    basic_report['name'] = app.config['NAME']
+    return render_template('overview.html', title='Overview', form=form,
+                            tables=tables, report=basic_report)
+
+@app.route('/earnings', methods=['GET', 'POST'])
+def earnings():
+    form = TestForm()
+
+    if (form.category_0.data):
+        flash("Voce selecionou {}".format(form.category_0.data))
+    if form.pay_method.data == 'Dinheiro':
+        flash("Voce selecionou {}".format(form.pay_method.data))
+    if form.submit.data:
+        flash("date: {}. Type: {}".format(form.date.data, type(form.date.data)))
+        flash("categoria: {}. Type: {}".format(form.category_0.data, type(form.category_0.data)))
+    return render_template('earnings.html', title='Test Zone', form=form)